rename advanceFilter service
