+++ src/services/pages/master-service.ts
@@ -2,7 +2,7 @@ import { Model, ModelFilter } from "@react3l/react3l/core";
 import { PRICE_LIST_DETAIL_ROUTE_PREFIX } from "config/route-consts";
 import { useCallback, useState } from "react";
 import { Observable } from "rxjs";
-import { advanceFilterService } from "services/AdvanceFilterService";
+import { advanceFilterService } from "services/advance-filter-service";
 import { queryStringService } from "services/QueryStringService";
 import { routerService } from "services/RouterService";
 import tableService from "services/table-service";
