+++ src/views/App/IndirectSalesOrderView/IndirectSalesOrderMasterView/IndirectSalesOrderMasterView.tsx
@@ -11,7 +11,7 @@ import AdvanceNumberRangeFilter from "components/Utility/AdvanceFilter/AdvanceNu
 import AdvanceStringFilter from "components/Utility/AdvanceFilter/AdvanceStringFilter/AdvanceStringFilter";
 import React from "react";
 import { Observable } from "rxjs";
-import { advanceFilterService } from "services/AdvanceFilterService";
+import { advanceFilterService } from "services/advance-filter-service";
 import { queryStringService } from "services/QueryStringService";
 
 class DemoFilter extends ModelFilter {
